requirements.txt
aiofiles
redis>=3.3.10
python-decouple==3.3
aiohttp[speedups]
cloudscraper
colour
covid
motor
bs4
cowpy
dataclasses
DateTime
emoji
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==4.0.0-rc1
gtts
hachoir
heroku3
html-telegraph-poster
humanize
jikanpy
justwatch
lottie
lyricsgenius
motor
moviepy
nekos.py
Pillow
prettytable
prsaw
psutil
psycopg2
pyfiglet
PyGithub
pygments
pylast
pymediainfo
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
regex
requests
search-engine-parser
selenium
setuptools
ShazamAPI
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph
ffmpeg-python
wheel
TgCrypto
tswift
ujson
urlextract
validators
pendulum
twitter_scraper
opencv-python
vcsi
wand
wget
wikipedia
telethon>=1.21.1
telethon-session-sqlalchemy
youtube-search-python==1.4.7
youtube_dl==2021.6.6
